#314
- modal: добавлены e2e-тесты
- reference.md: мелкие доработки
