+++ REFERENCE.md
@@ -15,8 +15,10 @@
 
 ### Доступность (Accessibility)
 
+По закону Парето для определения уровня доступности компонента можно ответить всего на пару вопросов:
+
 - Компонент хорошо виден пользователям с проблемами со зрением? (например фокус выделяется заметной рамкой)
-- Компонент реагирует на ввод с клавиатуры? (полноценная работа с компонентом должна быть возможна без мыши)
+- Компонент реагирует на ввод с клавиатуры? (полноценная работа с компонентом должна быть возможна на компьютере без мыши)
 
 ### Удобство в разработке (DX)
 
@@ -90,7 +92,7 @@ export function MyComponent({ size, className }: MyComponentProps) {
      height: 32px;
    }
 
-   @mixin size-s {
+   @mixin size-m {
      height: 48px;
    }
    ```
