+++ src/modal/__stories__/styles/additional-topbar.module.scss
